[TASK] Drop TYPO3 v8 and v9 support and cleanup deprecated code
