+++ .travis.yml
@@ -3,22 +3,24 @@ language: php
 matrix:
   fast_finish: true
   include:
-    - php: 7.1
-      env: TYPO3_VERSION=^7
-    - php: 7.1
-      env: TYPO3_VERSION=^8
-    - php: 7.2
-      env: TYPO3_VERSION=^9
-    - php: 7.3
-      env: TYPO3_VERSION=^9
-    - php: 7.4
-      env: TYPO3_VERSION=^9
     - php: 7.2
       env: TYPO3_VERSION=^10
     - php: 7.3
       env: TYPO3_VERSION=^10
     - php: 7.4
       env: TYPO3_VERSION=^10
+    - php: 7.4
+      env: TYPO3_VERSION=^11
+    - php: 8.0
+      env: TYPO3_VERSION=^11
+    - php: 8.1
+      env: TYPO3_VERSION=^11
+    - php: 8.2
+      env: TYPO3_VERSION=^11
+    - php: 8.1
+      env: TYPO3_VERSION=^12
+    - php: 8.2
+      env: TYPO3_VERSION=^12
 
 sudo: false
