+++ composer.json
@@ -25,11 +25,11 @@
         "typo3-ter/fal_s3": "self.version"
     },
     "require": {
+        "php": ">=7.2, <8.3",
         "aws/aws-sdk-php": "^3.199",
-        "php": ">=7.1, <8.2",
-        "typo3/cms-backend": "^8.7 || ^9.5 || ^10.4 || ^11.5",
-        "typo3/cms-core": "^8.7 || ^9.5 || ^10.4 || ^11.5",
-        "typo3/cms-extbase": "^8.7 || ^9.5 || ^10.4 || ^11.5"
+        "typo3/cms-backend": "^10.4 || ^11.5 || ^12.4",
+        "typo3/cms-core": "^10.4 || ^11.5 || ^12.4",
+        "typo3/cms-extbase": "^10.4 || ^11.5 || ^12.4"
     },
     "require-dev": {
         "nimut/testing-framework": "^5.1 || ^6.0"
