+++ ext_emconf.php
@@ -11,9 +11,9 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '1.14.2',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-11.5.99',
-            'backend' => '8.7.0-11.5.99',
-            'extbase' => '8.7.0-11.5.99',
+            'typo3' => '10.4.0-12.5.99',
+            'backend' => '10.4.0-12.5.99',
+            'extbase' => '10.4.0-12.5.99',
         ],
         'suggests' => [
             'extractor' => '2.0.0-2.99.99'
